+++ libraries/effect/src/main/java/androidx/media3/effect/BitmapTextureManager.java
@@ -21,7 +21,6 @@ import static androidx.media3.common.util.Assertions.checkState;
 
 import android.graphics.Bitmap;
 import androidx.media3.common.C;
-import androidx.media3.common.ColorInfo;
 import androidx.media3.common.FrameInfo;
 import androidx.media3.common.GlObjectsProvider;
 import androidx.media3.common.GlTextureInfo;
@@ -45,12 +44,11 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
   private final Queue<BitmapFrameSequenceInfo> pendingBitmaps;
   private final GlObjectsProvider glObjectsProvider;
 
-  private @MonotonicNonNull GlShaderProgram shaderProgram;
-  private @MonotonicNonNull GlTextureInfo currentGlTextureInfo;
+  private @MonotonicNonNull GainmapShaderProgram gainmapShaderProgram;
+  private @MonotonicNonNull GlTextureInfo currentSdrGlTextureInfo;
   private int downstreamShaderProgramCapacity;
   private boolean currentInputStreamEnded;
   private boolean isNextFrameInTexture;
-  private boolean useHdr;
 
   /**
    * Creates a new instance.
@@ -67,10 +65,16 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     pendingBitmaps = new LinkedBlockingQueue<>();
   }
 
+  /**
+   * {@inheritDoc}
+   *
+   * <p>{@link GlShaderProgram} must be a {@link GainmapShaderProgram}.
+   */
   @Override
   public void setSamplingGlShaderProgram(GlShaderProgram samplingGlShaderProgram) {
+    checkState(samplingGlShaderProgram instanceof GainmapShaderProgram);
     downstreamShaderProgramCapacity = 0;
-    this.shaderProgram = samplingGlShaderProgram;
+    this.gainmapShaderProgram = (GainmapShaderProgram) samplingGlShaderProgram;
   }
 
   @Override
@@ -93,11 +97,6 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
   }
 
   @Override
-  public void setInputFrameInfo(FrameInfo inputFrameInfo) {
-    this.useHdr = ColorInfo.isTransferHdr(inputFrameInfo.colorInfo);
-  }
-
-  @Override
   public int getPendingFrameCount() {
     // Always treat all queued bitmaps as immediately processed.
     return 0;
@@ -108,7 +107,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     videoFrameProcessingTaskExecutor.submit(
         () -> {
           if (pendingBitmaps.isEmpty()) {
-            checkNotNull(shaderProgram).signalEndOfCurrentInputStream();
+            checkNotNull(gainmapShaderProgram).signalEndOfCurrentInputStream();
             DebugTraceUtil.logEvent(
                 DebugTraceUtil.EVENT_BITMAP_TEXTURE_MANAGER_SIGNAL_EOS, C.TIME_END_OF_SOURCE);
           } else {
@@ -121,8 +120,8 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
   public void release() {
     videoFrameProcessingTaskExecutor.submit(
         () -> {
-          if (currentGlTextureInfo != null) {
-            currentGlTextureInfo.release();
+          if (currentSdrGlTextureInfo != null) {
+            currentSdrGlTextureInfo.release();
           }
           pendingBitmaps.clear();
         });
@@ -153,9 +152,9 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     }
 
     downstreamShaderProgramCapacity--;
-    checkNotNull(shaderProgram)
+    checkNotNull(gainmapShaderProgram)
         .queueInputFrame(
-            glObjectsProvider, checkNotNull(currentGlTextureInfo), currentPresentationTimeUs);
+            glObjectsProvider, checkNotNull(currentSdrGlTextureInfo), currentPresentationTimeUs);
     DebugTraceUtil.logEvent(
         DebugTraceUtil.EVENT_VFP_QUEUE_BITMAP,
         currentPresentationTimeUs,
@@ -169,7 +168,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
       finishedBitmapInfo.bitmap.recycle();
       if (pendingBitmaps.isEmpty() && currentInputStreamEnded) {
         // Only signal end of stream after all pending bitmaps are processed.
-        checkNotNull(shaderProgram).signalEndOfCurrentInputStream();
+        checkNotNull(gainmapShaderProgram).signalEndOfCurrentInputStream();
         DebugTraceUtil.logEvent(
             DebugTraceUtil.EVENT_BITMAP_TEXTURE_MANAGER_SIGNAL_EOS, C.TIME_END_OF_SOURCE);
         currentInputStreamEnded = false;
@@ -201,28 +200,19 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
       throws VideoFrameProcessingException {
     int currentTexId;
     try {
-      if (currentGlTextureInfo != null) {
-        currentGlTextureInfo.release();
+      if (currentSdrGlTextureInfo != null) {
+        currentSdrGlTextureInfo.release();
       }
       currentTexId = GlUtil.createTexture(bitmap);
-      if (useHdr && Util.SDK_INT >= 34 && bitmap.hasGainmap()) {
-        currentGlTextureInfo =
-            new GlTextureInfo(
-                currentTexId,
-                /* fboId= */ C.INDEX_UNSET,
-                /* rboId= */ C.INDEX_UNSET,
-                frameInfo.width,
-                frameInfo.height,
-                checkNotNull(bitmap.getGainmap()),
-                GlUtil.createTexture(bitmap.getGainmap().getGainmapContents()));
-      } else {
-        currentGlTextureInfo =
-            new GlTextureInfo(
-                currentTexId,
-                /* fboId= */ C.INDEX_UNSET,
-                /* rboId= */ C.INDEX_UNSET,
-                frameInfo.width,
-                frameInfo.height);
+      currentSdrGlTextureInfo =
+          new GlTextureInfo(
+              currentTexId,
+              /* fboId= */ C.INDEX_UNSET,
+              /* rboId= */ C.INDEX_UNSET,
+              frameInfo.width,
+              frameInfo.height);
+      if (Util.SDK_INT >= 34 && bitmap.hasGainmap()) {
+        checkNotNull(gainmapShaderProgram).setGainmap(checkNotNull(bitmap.getGainmap()));
       }
     } catch (GlUtil.GlException e) {
       throw VideoFrameProcessingException.from(e);