+++ libraries/transformer/src/androidTest/java/androidx/media3/transformer/mh/DefaultVideoFrameProcessorTextureOutputPixelTest.java
@@ -26,12 +26,14 @@ import static androidx.media3.transformer.AndroidTestUtil.MP4_ASSET_1080P_5_SECO
 import static androidx.media3.transformer.AndroidTestUtil.MP4_ASSET_720P_4_SECOND_HDR10_FORMAT;
 import static androidx.media3.transformer.AndroidTestUtil.MP4_ASSET_FORMAT;
 import static androidx.media3.transformer.AndroidTestUtil.recordTestSkipped;
+import static androidx.media3.transformer.EncoderUtil.getSupportedEncodersForHdrEditing;
 import static androidx.media3.transformer.mh.UnoptimizedGlEffect.NO_OP_EFFECT;
 import static androidx.test.core.app.ApplicationProvider.getApplicationContext;
 import static com.google.common.truth.Truth.assertThat;
 
 import android.content.Context;
 import android.graphics.Bitmap;
+import androidx.media3.common.C;
 import androidx.media3.common.ColorInfo;
 import androidx.media3.common.Effect;
 import androidx.media3.common.Format;
@@ -39,6 +41,7 @@ import androidx.media3.common.GlObjectsProvider;
 import androidx.media3.common.GlTextureInfo;
 import androidx.media3.common.VideoFrameProcessingException;
 import androidx.media3.common.util.GlUtil;
+import androidx.media3.common.util.Util;
 import androidx.media3.effect.BitmapOverlay;
 import androidx.media3.effect.DefaultGlObjectsProvider;
 import androidx.media3.effect.DefaultVideoFrameProcessor;
@@ -48,12 +51,14 @@ import androidx.media3.test.utils.BitmapPixelTestUtil;
 import androidx.media3.test.utils.TextureBitmapReader;
 import androidx.media3.test.utils.VideoFrameProcessorTestRunner;
 import androidx.media3.transformer.AndroidTestUtil;
-import androidx.media3.transformer.EncoderUtil;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import com.google.common.collect.ImmutableList;
+import java.io.IOException;
 import java.util.List;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
+import org.json.JSONException;
 import org.junit.After;
+import org.junit.AssumptionViolatedException;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -75,11 +80,16 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
   private static final String BITMAP_OVERLAY_PNG_ASSET_PATH =
       "test-generated-goldens/sample_mp4_first_frame/electrical_colors/overlay_bitmap_FrameProcessor.png";
   private static final String OVERLAY_PNG_ASSET_PATH = "media/png/media3test.png";
+  private static final String ULTRA_HDR_ASSET_PATH = "media/jpeg/ultraHDR.jpg";
 
   private static final String ORIGINAL_HLG10_PNG_ASSET_PATH =
       "test-generated-goldens/sample_mp4_first_frame/electrical_colors/original_hlg10.png";
   private static final String ORIGINAL_HDR10_PNG_ASSET_PATH =
       "test-generated-goldens/sample_mp4_first_frame/electrical_colors/original_hdr10.png";
+  private static final String ULTRA_HDR_TO_HLG_PNG_ASSET_PATH =
+      "test-generated-goldens/hdr-goldens/ultrahdr_to_hlg.png";
+  private static final String ULTRA_HDR_TO_PQ_PNG_ASSET_PATH =
+      "test-generated-goldens/hdr-goldens/ultrahdr_to_pq.png";
 
   /** Input SDR video of which we only use the first frame. */
   private static final String INPUT_SDR_MP4_ASSET_STRING = "media/mp4/sample.mp4";
@@ -117,7 +127,7 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
         /* outputFormat= */ null)) {
       return;
     }
-    videoFrameProcessorTestRunner = getSurfaceInputFrameProcessorTestRunnerBuilder(testId).build();
+    videoFrameProcessorTestRunner = getDefaultFrameProcessorTestRunnerBuilder(testId).build();
     Bitmap expectedBitmap = readBitmap(ORIGINAL_PNG_ASSET_PATH);
 
     videoFrameProcessorTestRunner.processFirstFrameAndEnd();
@@ -172,7 +182,7 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
     Bitmap overlayBitmap = readBitmap(OVERLAY_PNG_ASSET_PATH);
     BitmapOverlay bitmapOverlay = BitmapOverlay.createStaticBitmapOverlay(overlayBitmap);
     videoFrameProcessorTestRunner =
-        getSurfaceInputFrameProcessorTestRunnerBuilder(testId)
+        getDefaultFrameProcessorTestRunnerBuilder(testId)
             .setEffects(new OverlayEffect(ImmutableList.of(bitmapOverlay)))
             .build();
     Bitmap expectedBitmap = readBitmap(BITMAP_OVERLAY_PNG_ASSET_PATH);
@@ -220,17 +230,14 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
   public void noEffects_hlg10Input_matchesGoldenFile() throws Exception {
     Context context = getApplicationContext();
     Format format = MP4_ASSET_1080P_5_SECOND_HLG10_FORMAT;
-    if (!deviceSupportsHdrEditing(format)) {
-      recordTestSkipped(context, testId, "No HLG editing support");
-      return;
-    }
+    assumeDeviceSupportsHdrEditing(format);
     if (AndroidTestUtil.skipAndLogIfFormatsUnsupported(
         context, testId, /* inputFormat= */ format, /* outputFormat= */ null)) {
       return;
     }
     ColorInfo colorInfo = checkNotNull(format.colorInfo);
     videoFrameProcessorTestRunner =
-        getSurfaceInputFrameProcessorTestRunnerBuilder(testId)
+        getDefaultFrameProcessorTestRunnerBuilder(testId)
             .setOutputColorInfo(colorInfo)
             .setVideoAssetPath(INPUT_HLG10_MP4_ASSET_STRING)
             .build();
@@ -251,10 +258,7 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
   public void noEffects_hlg10TextureInput_matchesGoldenFile() throws Exception {
     Context context = getApplicationContext();
     Format format = MP4_ASSET_1080P_5_SECOND_HLG10_FORMAT;
-    if (!deviceSupportsHdrEditing(format)) {
-      recordTestSkipped(context, testId, "No HLG editing support");
-      return;
-    }
+    assumeDeviceSupportsHdrEditing(format);
     if (AndroidTestUtil.skipAndLogIfFormatsUnsupported(
         context, testId, /* inputFormat= */ format, /* outputFormat= */ null)) {
       return;
@@ -283,20 +287,50 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
   }
 
   @Test
+  public void noEffects_hlg10UltraHDRImageInput_matchesGoldenFile() throws Exception {
+    assumeDeviceSupportsHdrEditing(MP4_ASSET_1080P_5_SECOND_HLG10_FORMAT);
+    assumeDeviceSupportsUltraHdrEditing();
+    ColorInfo outputColorInfo =
+        new ColorInfo.Builder()
+            .setColorSpace(C.COLOR_SPACE_BT2020)
+            .setColorTransfer(C.COLOR_TRANSFER_HLG)
+            .setColorRange(C.COLOR_RANGE_FULL)
+            .build();
+    videoFrameProcessorTestRunner =
+        getDefaultFrameProcessorTestRunnerBuilder(testId)
+            .setOutputColorInfo(outputColorInfo)
+            .build();
+    Bitmap originalBitmap = readBitmap(ULTRA_HDR_ASSET_PATH);
+    Bitmap expectedBitmap = readBitmap(ULTRA_HDR_TO_HLG_PNG_ASSET_PATH);
+
+    videoFrameProcessorTestRunner.queueInputBitmap(
+        originalBitmap,
+        /* durationUs= */ C.MICROS_PER_SECOND,
+        /* offsetToAddUs= */ 0L,
+        /* frameRate= */ 1);
+    videoFrameProcessorTestRunner.endFrameProcessing();
+    Bitmap actualBitmap = videoFrameProcessorTestRunner.getOutputBitmap();
+
+    // TODO(b/207848601): Switch to using proper tooling for testing against golden data.
+    float averagePixelAbsoluteDifference =
+        BitmapPixelTestUtil.getBitmapAveragePixelAbsoluteDifferenceFp16(
+            expectedBitmap, actualBitmap);
+    assertThat(averagePixelAbsoluteDifference)
+        .isAtMost(MAXIMUM_AVERAGE_PIXEL_ABSOLUTE_DIFFERENCE_DIFFERENT_DEVICE_FP16);
+  }
+
+  @Test
   public void noEffects_hdr10Input_matchesGoldenFile() throws Exception {
     Context context = getApplicationContext();
     Format format = MP4_ASSET_720P_4_SECOND_HDR10_FORMAT;
-    if (!deviceSupportsHdrEditing(format)) {
-      recordTestSkipped(context, testId, "No PQ editing support");
-      return;
-    }
+    assumeDeviceSupportsHdrEditing(format);
     if (AndroidTestUtil.skipAndLogIfFormatsUnsupported(
         context, testId, /* inputFormat= */ format, /* outputFormat= */ null)) {
       return;
     }
     ColorInfo colorInfo = checkNotNull(format.colorInfo);
     videoFrameProcessorTestRunner =
-        getSurfaceInputFrameProcessorTestRunnerBuilder(testId)
+        getDefaultFrameProcessorTestRunnerBuilder(testId)
             .setOutputColorInfo(colorInfo)
             .setVideoAssetPath(INPUT_PQ_MP4_ASSET_STRING)
             .build();
@@ -317,10 +351,7 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
   public void noEffects_hdr10TextureInput_matchesGoldenFile() throws Exception {
     Context context = getApplicationContext();
     Format format = MP4_ASSET_720P_4_SECOND_HDR10_FORMAT;
-    if (!deviceSupportsHdrEditing(format)) {
-      recordTestSkipped(context, testId, "No PQ editing support");
-      return;
-    }
+    assumeDeviceSupportsHdrEditing(format);
     if (AndroidTestUtil.skipAndLogIfFormatsUnsupported(
         context, testId, /* inputFormat= */ format, /* outputFormat= */ null)) {
       return;
@@ -349,20 +380,50 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
   }
 
   @Test
+  public void noEffects_hdr10UltraHDRImageInput_matchesGoldenFile() throws Exception {
+    assumeDeviceSupportsHdrEditing(MP4_ASSET_720P_4_SECOND_HDR10_FORMAT);
+    assumeDeviceSupportsUltraHdrEditing();
+    ColorInfo outputColorInfo =
+        new ColorInfo.Builder()
+            .setColorSpace(C.COLOR_SPACE_BT2020)
+            .setColorTransfer(C.COLOR_TRANSFER_ST2084)
+            .setColorRange(C.COLOR_RANGE_FULL)
+            .build();
+    videoFrameProcessorTestRunner =
+        getDefaultFrameProcessorTestRunnerBuilder(testId)
+            .setOutputColorInfo(outputColorInfo)
+            .build();
+    Bitmap originalBitmap = readBitmap(ULTRA_HDR_ASSET_PATH);
+    Bitmap expectedBitmap = readBitmap(ULTRA_HDR_TO_PQ_PNG_ASSET_PATH);
+
+    videoFrameProcessorTestRunner.queueInputBitmap(
+        originalBitmap,
+        /* durationUs= */ C.MICROS_PER_SECOND,
+        /* offsetToAddUs= */ 0L,
+        /* frameRate= */ 1);
+    videoFrameProcessorTestRunner.endFrameProcessing();
+    Bitmap actualBitmap = videoFrameProcessorTestRunner.getOutputBitmap();
+
+    // TODO(b/207848601): Switch to using proper tooling for testing against golden data.
+    float averagePixelAbsoluteDifference =
+        BitmapPixelTestUtil.getBitmapAveragePixelAbsoluteDifferenceFp16(
+            expectedBitmap, actualBitmap);
+    assertThat(averagePixelAbsoluteDifference)
+        .isAtMost(MAXIMUM_AVERAGE_PIXEL_ABSOLUTE_DIFFERENCE_DIFFERENT_DEVICE_FP16);
+  }
+
+  @Test
   public void noOpEffect_hlg10Input_matchesGoldenFile() throws Exception {
     Context context = getApplicationContext();
     Format format = MP4_ASSET_1080P_5_SECOND_HLG10_FORMAT;
-    if (!deviceSupportsHdrEditing(format)) {
-      recordTestSkipped(context, testId, "No HLG editing support");
-      return;
-    }
+    assumeDeviceSupportsHdrEditing(format);
     if (AndroidTestUtil.skipAndLogIfFormatsUnsupported(
         context, testId, /* inputFormat= */ format, /* outputFormat= */ null)) {
       return;
     }
     ColorInfo colorInfo = checkNotNull(format.colorInfo);
     videoFrameProcessorTestRunner =
-        getSurfaceInputFrameProcessorTestRunnerBuilder(testId)
+        getDefaultFrameProcessorTestRunnerBuilder(testId)
             .setOutputColorInfo(colorInfo)
             .setVideoAssetPath(INPUT_HLG10_MP4_ASSET_STRING)
             .setEffects(NO_OP_EFFECT)
@@ -384,10 +445,7 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
   public void noOpEffect_hlg10TextureInput_matchesGoldenFile() throws Exception {
     Context context = getApplicationContext();
     Format format = MP4_ASSET_1080P_5_SECOND_HLG10_FORMAT;
-    if (!deviceSupportsHdrEditing(format)) {
-      recordTestSkipped(context, testId, "No HLG editing support");
-      return;
-    }
+    assumeDeviceSupportsHdrEditing(format);
     if (AndroidTestUtil.skipAndLogIfFormatsUnsupported(
         context, testId, /* inputFormat= */ format, /* outputFormat= */ null)) {
       return;
@@ -416,20 +474,51 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
   }
 
   @Test
+  public void noOpEffect_hlg10UltraHDRImageInput_matchesGoldenFile() throws Exception {
+    assumeDeviceSupportsHdrEditing(MP4_ASSET_720P_4_SECOND_HDR10_FORMAT);
+    assumeDeviceSupportsUltraHdrEditing();
+    ColorInfo outputColorInfo =
+        new ColorInfo.Builder()
+            .setColorSpace(C.COLOR_SPACE_BT2020)
+            .setColorTransfer(C.COLOR_TRANSFER_HLG)
+            .setColorRange(C.COLOR_RANGE_FULL)
+            .build();
+    videoFrameProcessorTestRunner =
+        getDefaultFrameProcessorTestRunnerBuilder(testId)
+            .setOutputColorInfo(outputColorInfo)
+            .setEffects(NO_OP_EFFECT)
+            .build();
+    Bitmap originalBitmap = readBitmap(ULTRA_HDR_ASSET_PATH);
+    Bitmap expectedBitmap = readBitmap(ULTRA_HDR_TO_HLG_PNG_ASSET_PATH);
+
+    videoFrameProcessorTestRunner.queueInputBitmap(
+        originalBitmap,
+        /* durationUs= */ C.MICROS_PER_SECOND,
+        /* offsetToAddUs= */ 0L,
+        /* frameRate= */ 1);
+    videoFrameProcessorTestRunner.endFrameProcessing();
+    Bitmap actualBitmap = videoFrameProcessorTestRunner.getOutputBitmap();
+
+    // TODO(b/207848601): Switch to using proper tooling for testing against golden data.
+    float averagePixelAbsoluteDifference =
+        BitmapPixelTestUtil.getBitmapAveragePixelAbsoluteDifferenceFp16(
+            expectedBitmap, actualBitmap);
+    assertThat(averagePixelAbsoluteDifference)
+        .isAtMost(MAXIMUM_AVERAGE_PIXEL_ABSOLUTE_DIFFERENCE_DIFFERENT_DEVICE_FP16);
+  }
+
+  @Test
   public void noOpEffect_hdr10Input_matchesGoldenFile() throws Exception {
     Context context = getApplicationContext();
     Format format = MP4_ASSET_720P_4_SECOND_HDR10_FORMAT;
-    if (!deviceSupportsHdrEditing(format)) {
-      recordTestSkipped(context, testId, "No PQ editing support");
-      return;
-    }
+    assumeDeviceSupportsHdrEditing(format);
     if (AndroidTestUtil.skipAndLogIfFormatsUnsupported(
         context, testId, /* inputFormat= */ format, /* outputFormat= */ null)) {
       return;
     }
     ColorInfo colorInfo = checkNotNull(format.colorInfo);
     videoFrameProcessorTestRunner =
-        getSurfaceInputFrameProcessorTestRunnerBuilder(testId)
+        getDefaultFrameProcessorTestRunnerBuilder(testId)
             .setOutputColorInfo(colorInfo)
             .setVideoAssetPath(INPUT_PQ_MP4_ASSET_STRING)
             .setEffects(NO_OP_EFFECT)
@@ -451,10 +540,7 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
   public void noOpEffect_hdr10TextureInput_matchesGoldenFile() throws Exception {
     Context context = getApplicationContext();
     Format format = MP4_ASSET_720P_4_SECOND_HDR10_FORMAT;
-    if (!deviceSupportsHdrEditing(format)) {
-      recordTestSkipped(context, testId, "No PQ editing support");
-      return;
-    }
+    assumeDeviceSupportsHdrEditing(format);
     if (AndroidTestUtil.skipAndLogIfFormatsUnsupported(
         context, testId, /* inputFormat= */ format, /* outputFormat= */ null)) {
       return;
@@ -473,6 +559,40 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
     texIdProducingVideoFrameProcessorTestRunner.processFirstFrameAndEnd();
     texIdProducingVideoFrameProcessorTestRunner.release();
     Bitmap actualBitmap = consumersBitmapReader.getBitmap();
+
+    // TODO(b/207848601): Switch to using proper tooling for testing against golden data.
+    float averagePixelAbsoluteDifference =
+        BitmapPixelTestUtil.getBitmapAveragePixelAbsoluteDifferenceFp16(
+            expectedBitmap, actualBitmap);
+    assertThat(averagePixelAbsoluteDifference)
+        .isAtMost(MAXIMUM_AVERAGE_PIXEL_ABSOLUTE_DIFFERENCE_DIFFERENT_DEVICE_FP16);
+  }
+
+  @Test
+  public void noOpEffect_hdr10UltraHDRImageInput_matchesGoldenFile() throws Exception {
+    assumeDeviceSupportsHdrEditing(MP4_ASSET_720P_4_SECOND_HDR10_FORMAT);
+    assumeDeviceSupportsUltraHdrEditing();
+    ColorInfo outputColorInfo =
+        new ColorInfo.Builder()
+            .setColorSpace(C.COLOR_SPACE_BT2020)
+            .setColorTransfer(C.COLOR_TRANSFER_ST2084)
+            .setColorRange(C.COLOR_RANGE_FULL)
+            .build();
+    videoFrameProcessorTestRunner =
+        getDefaultFrameProcessorTestRunnerBuilder(testId)
+            .setOutputColorInfo(outputColorInfo)
+            .setEffects(NO_OP_EFFECT)
+            .build();
+    Bitmap originalBitmap = readBitmap(ULTRA_HDR_ASSET_PATH);
+    Bitmap expectedBitmap = readBitmap(ULTRA_HDR_TO_PQ_PNG_ASSET_PATH);
+
+    videoFrameProcessorTestRunner.queueInputBitmap(
+        originalBitmap,
+        /* durationUs= */ C.MICROS_PER_SECOND,
+        /* offsetToAddUs= */ 0L,
+        /* frameRate= */ 1);
+    videoFrameProcessorTestRunner.endFrameProcessing();
+    Bitmap actualBitmap = videoFrameProcessorTestRunner.getOutputBitmap();
 
     // TODO(b/207848601): Switch to using proper tooling for testing against golden data.
     float averagePixelAbsoluteDifference =
@@ -555,7 +675,7 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
     textureProducer.releaseOutputTexture(presentationTimeUs);
   }
 
-  private VideoFrameProcessorTestRunner.Builder getSurfaceInputFrameProcessorTestRunnerBuilder(
+  private VideoFrameProcessorTestRunner.Builder getDefaultFrameProcessorTestRunnerBuilder(
       String testId) {
     TextureBitmapReader textureBitmapReader = new TextureBitmapReader();
     DefaultVideoFrameProcessor.Factory defaultVideoFrameProcessorFactory =
@@ -574,9 +694,22 @@ public final class DefaultVideoFrameProcessorTextureOutputPixelTest {
         .setBitmapReader(textureBitmapReader);
   }
 
-  private static boolean deviceSupportsHdrEditing(Format format) {
-    return !EncoderUtil.getSupportedEncodersForHdrEditing(
-            checkNotNull(checkNotNull(format).sampleMimeType), format.colorInfo)
-        .isEmpty();
+  private void assumeDeviceSupportsHdrEditing(Format format) throws JSONException, IOException {
+    if (getSupportedEncodersForHdrEditing(format.sampleMimeType, format.colorInfo).isEmpty()) {
+      String skipReason =
+          format.colorInfo.colorTransfer == C.COLOR_TRANSFER_HLG
+              ? "No HLG editing support."
+              : "No PQ editing support.";
+      recordTestSkipped(getApplicationContext(), testId, skipReason);
+      throw new AssumptionViolatedException(skipReason);
+    }
+  }
+
+  private void assumeDeviceSupportsUltraHdrEditing() throws JSONException, IOException {
+    if (Util.SDK_INT < 34) {
+      recordTestSkipped(
+          getApplicationContext(), testId, "Ultra HDR is not supported on this API level.");
+      throw new AssumptionViolatedException("Ultra HDR is not supported on this API level.");
+    }
   }
 }